+++ R/00-Namespace.R
@@ -10,7 +10,8 @@ utils::globalVariables(
 #' @importFrom httr2 request req_perform req_timeout req_user_agent resp_status resp_body_html
 #' @importFrom jsonlite fromJSON toJSON write_json
 #' @importFrom rvest html_element html_elements html_text2
-#' @importFrom shinyTools shinyTryCatch
+#' @importFrom shinyTools headerButtonsUI shinyTryCatch
+#' @importFrom shinyTree renderTree shinyTree
 #' @importFrom stats setNames
 #' @importFrom stringr str_squish
 #' @importFrom tibble tibble