+++ tests/testthat/test-example-workflow.R
@@ -39,7 +39,7 @@ test_that("workflow can be exported, imported, and run as in example_wf.R", {
   # Run imported workflow (steps 1 to 2)
   my_run_2 <- PEITHO::run(my_wf_imported, from = 1, to = 2)
   expect_type(my_run_2$state$last_result, "list")
-  expect_length(my_run_2$state$last_result, 3)
+  expect_length(my_run_2$state$last_result, 2)
 
   # Truncate output
   trunc_out2 <- PEITHO:::trunc(my_run_2$state$last_result, n_char = 100)